oopsies yellow LED functionality
red LED and buzzer toggles when any charger fault is communicated by the BMS. yellow LED toggles when there is no error, BMS indicates charger is ready, but the shutdown circuit is open. green LED toggles when there is no error, BMS indicates charger is ready, and the shutdown circuit is closed.